refactor(attendance): split capture policy helpers from service
